_exposureMode and _focusMode aren't initialized
These properties aren't initialized with real default values of _currentDevice property. So, it causes a bug because these take the default values as 0 in [PBJVision sharedInstance], it means 

_exposureMode   = PBJExposureModeLocked;
_focusMode      = PBJFocusModeLocked;

then, if you try

PBJVision *vision = [PBJVision sharedInstance];
vision.focusMode    = PBJFocusModeLocked;
vision.exposureMode = PBJExposureModeLocked;

it never executes lines 469-477 and 493-499 respectively, because the line 465 and 487 always is `false`. But the real values of _currentDevice.focusMode and _currentDevice.exposureMode are AVCaptureFocusModeContinuousAutoFocus and AVCaptureExposureModeContinuousAutoExposure.